requirements.txt
gspread~=4.0.1
requests~=2.26.0
python-dotenv~=0.19.0
pyTelegramBotAPI~=4.1.0